Filestore: Only verify contents of block if the file's mod-time has changed.
In DataObj add two fields: the file's modification time, and a flag to
mark the content as invalid.  If the modification time of the file is
the same as the one in the DataObj, skip validation and return the
cached value of the invalid flag.  If the modification time differs
than recheck the hash of the block and update the modification time
and invalid flag accordingly.

With this change retrieval of blocks from the filestore is as fast as
retrieving a block from the normal datastore.

Note that when a block is first added the modification time is unset
so the first retrial will be slower since the contents need to be
verified.  After the first retrieval no additional verification needs
to be done until the file changed.

License: MIT
Signed-off-by: Kevin Atkinson <k@kevina.org>