Filestore: Get the modification time from the file before it is first read.
When adding a file to the filestore get the modification time of a
file before we read any data from it and use that time to set the
initial timestamp in the DataObj.  This avoids the need for verifying a
file when it is retrieved for the first time.

License: MIT
Signed-off-by: Kevin Atkinson <k@kevina.org>