Redo "filestore verify" to check more than if blocks can be read.
In addition to checking if blocks can be read also check for other
problems.  (1) For each block representing a file root, check that all
the blocks in the merkledag tree can be read and the contents of the
file can be reconstructed.  (2) Check for orphan nodes, those are
nodes that are a child of another node that was not found in the
filestore. (3) Check if a file was appended too as the block is still
valid but no longer represents the contents of the file.

The "missing" status, which was used indicate that the file is not
found, is now "no-file" and "missing" is now used to indicate when a
block could not be found in the datastore.

The original verify behavior is still available by using the "--basic"
option.

License: MIT
Signed-off-by: Kevin Atkinson <k@kevina.org>