Add "filestore fix-pins", remove obsolete "filestore find-dangling-pins".
The new command will repair broken pins by either removing or
reconstructing the pin so that invalid blocks are not pinned.  If the
pin points to a block that no longer exists it is simply removed.  If
the pin is a recursive pin that points to a valid block but some of
the nodes of the merkle tree are invalid the recursive pin is
converted to a direct pin and new recursive pins are created for the
nodes that are still valid.

License: MIT
Signed-off-by: Kevin Atkinson <k@kevina.org>