Create a new AdvReader interface.
Create a new AdvReader interface that returns an ExtraInfo object that
contains extra information about the the filehandle.  This includes
the current offset and the absolute path of a file.

Also modify chunk.Splitter to return a Bytes struct in the NextBytes()
method that in addition to the raw bytes returns the ExtraInfo object
from the AdvReader that contains the current offset in the file
(amoung other information).

License: MIT
Signed-off-by: Kevin Atkinson <k@kevina.org>