Replace the multi-datastore with a multi-blockstore.
Instead of using a multi-datastore, mount each datastore under a
different mount point and use a multi-blockstore to check each mount
point for the block.

The first mount checked of the multi-blockstore is considered the
"cache", all others are considered read-only.  This implies that the
garbage collector only removes block from the first mount.

This change also factors out the pinlock from the blockstore into its
own structure.  Only the multi-datastore now implements the
GCBlockstore interface.  In the future this could be separated out
from the blockstore completely.

For now caching is only done on the first mount, in the future this
could be reworked.  The bloom filter is the most problematic as the
read-only mounts are not necessary immutable and can be changed by
methods outside of the blockstore.

License: MIT
Signed-off-by: Kevin Atkinson <k@kevina.org>